src/requirements.txt
gpiozero==2.0
RPi.GPIO==0.7.1
google-cloud-speech==2.24.1
